replace old LowerTypes pass with NewLowerTypes pass
