+++ src/main/scala/firrtl/stage/Forms.scala
@@ -60,7 +60,7 @@ object Forms {
          Dependency[firrtl.transforms.formal.AssertSubmoduleAssumptions] )
 
   val LowForm: Seq[TransformDependency] = MidForm ++
-    Seq( Dependency(passes.NewLowerTypes),
+    Seq( Dependency(passes.LowerTypes),
          Dependency(passes.Legalize),
          Dependency(firrtl.transforms.RemoveReset),
          Dependency[firrtl.transforms.CheckCombLoops],
