+++ src/main/scala/firrtl/transforms/CheckCombLoops.scala
@@ -103,7 +103,7 @@ class CheckCombLoops extends Transform
     with DependencyAPIMigration {
 
   override def prerequisites = firrtl.stage.Forms.MidForm ++
-    Seq( Dependency(passes.NewLowerTypes),
+    Seq( Dependency(passes.LowerTypes),
          Dependency(passes.Legalize),
          Dependency(firrtl.transforms.RemoveReset) )
