+++ src/main/scala/firrtl/transforms/RemoveReset.scala
@@ -18,7 +18,7 @@ import scala.collection.{immutable, mutable}
 object RemoveReset extends Transform with DependencyAPIMigration {
 
   override def prerequisites = firrtl.stage.Forms.MidForm ++
-    Seq( Dependency(passes.NewLowerTypes),
+    Seq( Dependency(passes.LowerTypes),
          Dependency(passes.Legalize) )
 
   override def optionalPrerequisites = Seq.empty
