+++ src/main/scala/firrtl/transforms/RemoveWires.scala
@@ -23,7 +23,7 @@ import scala.util.{Try, Success, Failure}
 class RemoveWires extends Transform with DependencyAPIMigration {
 
   override def prerequisites = firrtl.stage.Forms.MidForm ++
-    Seq( Dependency(passes.NewLowerTypes),
+    Seq( Dependency(passes.LowerTypes),
          Dependency(passes.Legalize),
          Dependency(passes.ResolveKinds),
          Dependency(transforms.RemoveReset),
