+++ src/test/scala/firrtl/passes/LowerTypesSpec.scala
@@ -8,27 +8,19 @@ import firrtl.stage.TransformManager
 import firrtl.stage.TransformManager.TransformDependency
 import org.scalatest.flatspec.AnyFlatSpec
 
-class LegacyLowerTypesSpec extends LowerTypesEndToEndSpec {
-  override protected def pass: TransformDependency = Dependency(NewLowerTypes)
-}
 
-class NewLowerTypesEndToEndSpec extends LowerTypesEndToEndSpec {
-  override protected def pass: TransformDependency = Dependency(NewLowerTypes)
-}
-
-/** Unit test style tests for [[NewLowerTypes]].
+/** Unit test style tests for [[LowerTypes]].
   * You can find additional integration style tests in [[firrtlTests.LowerTypesSpec]]
   */
-class NewLowerTypesUnitTestSpec extends LowerTypesBaseSpec {
+class LowerTypesUnitTestSpec extends LowerTypesBaseSpec {
   import LowerTypesSpecUtils._
   override protected def lower(n: String, tpe: String, namespace: Set[String]): Seq[String] =
     destruct(n, tpe, namespace).fields
 }
 
 /** Runs the lowering pass in the context of the compiler instead of directly calling internal functions. */
-abstract class LowerTypesEndToEndSpec extends LowerTypesBaseSpec {
-  protected def pass: TransformDependency
-  private lazy val lowerTypesCompiler = new TransformManager(Seq(pass))
+class LowerTypesEndToEndSpec extends LowerTypesBaseSpec {
+  private lazy val lowerTypesCompiler = new TransformManager(Seq(Dependency(LowerTypes)))
   private def legacyLower(n: String, tpe: String, namespace: Set[String]): Seq[String] = {
     val inputs = namespace.map(n => s"    input $n : UInt<1>").mkString("\n")
     val src =
